emc: move the interpreter and motion interface types into nml_intf
Four includes in emc/nml_intf reached up into emc/rs274ngc, and one in
emc/rs274ngc reached into emc/motion, for types that neither of those
directories owns.

state_tag_t is filled in by the interpreter and carried in motion messages
and in TC_STRUCT; its own header says the info "isn't actually used by
motion directly".  StateTag is its C++ form, embedded in canon.hh, emc.hh
and emc_nml.hh, and emc/nml_intf/Submakefile already compiled modal_state.cc
into libemc from across the tree.  Both move to emc/nml_intf, which is where
the rest of the shared interface already lives.

emc_nml.hh included rs274ngc.hh, and so the whole interpreter class
hierarchy, for three array sizes.  ACTIVE_G_CODES, ACTIVE_M_CODES and
ACTIVE_SETTINGS move into emc/nml_intf/interp_codes.h, included by
interp_base.hh where they were and by emc_nml.hh where they are used.

Two includes go with that.  modal_state.cc included interp_base.hh and used
nothing from it.  emcmodule.cc was reaching debugflags.h through the
rs274ngc.hh chain and now names it.

emc/nml_intf no longer depends on emc/rs274ngc and emc/rs274ngc no longer
depends on emc/motion, which takes emc/rs274ngc, emc/ini and
emc/pythonplugin out of the cycle.

Twelve libc includes go in with it.  emc_nml.hh no longer pulls in the
interpreter header chain, and eight files under emc/ were reaching stdio.h,
stdlib.h and unistd.h through it rather than naming them.  On Debian sid that
chain does not lead there and getpid() in shcom.cc and emcsvr.cc does not
compile; on bookworm and trixie it still does, which is why one package job
caught it and the rest did not.
