+++ src/emc/nml_intf/Submakefile
@@ -1,7 +1,7 @@
 
 LIBEMCSRCS := \
     emc/nml_intf/emcglb.c \
-    emc/rs274ngc/modal_state.cc \
+    emc/nml_intf/modal_state.cc \
     emc/nml_intf/emc.cc \
     emc/nml_intf/emcpose.c \
     emc/nml_intf/emcargs.cc \
