+++ src/emc/nml_intf/canon.hh
@@ -20,7 +20,7 @@
 #include "emctool.h"
 #include "canon_position.hh"
 #include <emcmotcfg.h> // Just for EMCMOT_NUM_SPINDLES
-#include "rs274ngc/modal_state.hh"
+#include "modal_state.hh"
 
 /*
   This is the header file that all applications that use the
