+++ src/emc/nml_intf/interp_codes.h
@@ -1,0 +1,23 @@
+#ifndef __LINUXCNC_INTERP_CODES_H
+#define __LINUXCNC_INTERP_CODES_H
+//    This program is free software; you can redistribute it and/or modify
+//    it under the terms of the GNU General Public License as published by
+//    the Free Software Foundation; either version 2 of the License, or
+//    (at your option) any later version.
+//
+//    This program is distributed in the hope that it will be useful,
+//    but WITHOUT ANY WARRANTY; without even the implied warranty of
+//    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+//    GNU General Public License for more details.
+//
+//    You should have received a copy of the GNU General Public License
+//    along with this program; if not, write to the Free Software
+//    Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+
+/* Sizes of the arrays of active codes that the interpreter reports and
+   the status message carries. */
+#define ACTIVE_G_CODES 17
+#define ACTIVE_M_CODES 10
+#define ACTIVE_SETTINGS 5
+
+#endif
