+++ src/emc/nml_intf/modal_state.cc
@@ -19,7 +19,6 @@
 * along with this program; if not, write to the Free Software
 * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 ********************************************************************/
-#include "interp_base.hh"
 #include "modal_state.hh"
 #include <string.h>
