+++ src/emc/sai/dummyemcstat.cc
@@ -17,6 +17,7 @@
  */
 // keep linker happy so TaskMod can be resolved
 
+#include <stdio.h>
 #include "libnml/rcs/rcs.hh"		// NML classes, nmlErrorFormat()
 #include "nml_intf/emc.hh"		// EMC NML
 #include "nml_intf/emc_nml.hh"
