+++ src/emc/tooldata/tool_watch.cc
@@ -19,6 +19,7 @@
 ** along with this program; if not, write to the Free Software
 ** Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 */
+#include <stdio.h>
 #include "nml_intf/emc.hh"
 #include "nml_intf/emc_nml.hh"
 #include <unistd.h>
